requirements.txt
coloredlogs
pyaes
pyyaml